+++ examples/iam-user/outputs.tf
@@ -85,3 +85,8 @@ output "keybase_secret_key_pgp_message" {
   description = "Encrypted access secret key"
   value       = module.iam_user.keybase_secret_key_pgp_message
 }
+
+output "policy_arns" {
+  description = "The list of ARNs of policies directly assigned to the IAM user"
+  value       = module.iam_user.policy_arns
+}