+++ modules/iam-role-for-service-accounts-eks/variables.tf
@@ -86,6 +86,13 @@ variable "allow_self_assume_role" {
 # Policies
 ################################################################################
 
+# AWS Gateway Controller
+variable "attach_aws_gateway_controller_policy" {
+  description = "Determines whether to attach the AWS Gateway Controller IAM policy to the role"
+  type        = bool
+  default     = false
+}
+
 # Cert Manager
 variable "attach_cert_manager_policy" {
   description = "Determines whether to attach the Cert Manager IAM policy to the role"
@@ -107,7 +114,13 @@ variable "attach_cluster_autoscaler_policy" {
 }
 
 variable "cluster_autoscaler_cluster_ids" {
-  description = "List of cluster IDs to appropriately scope permissions within the Cluster Autoscaler IAM policy"
+  description = "[Deprecated - use `cluster_autoscaler_cluster_names`] List of cluster names to appropriately scope permissions within the Cluster Autoscaler IAM policy"
+  type        = list(string)
+  default     = []
+}
+
+variable "cluster_autoscaler_cluster_names" {
+  description = "List of cluster names to appropriately scope permissions within the Cluster Autoscaler IAM policy"
   type        = list(string)
   default     = []
 }
@@ -164,6 +177,18 @@ variable "external_secrets_secrets_manager_arns" {
   default     = ["arn:aws:secretsmanager:*:*:secret:*"]
 }
 
+variable "external_secrets_kms_key_arns" {
+  description = "List of KMS Key ARNs that are used by Secrets Manager that contain secrets to mount using External Secrets"
+  type        = list(string)
+  default     = ["arn:aws:kms:*:*:key/*"]
+}
+
+variable "external_secrets_secrets_manager_create_permission" {
+  description = "Determins whether External Secrets may use secretsmanager:CreateSecret"
+  type        = bool
+  default     = false
+}
+
 # FSx Lustre CSI
 variable "attach_fsx_lustre_csi_policy" {
   description = "Determines whether to attach the FSx for Lustre CSI Driver IAM policy to the role"
@@ -185,7 +210,13 @@ variable "attach_karpenter_controller_policy" {
 }
 
 variable "karpenter_controller_cluster_id" {
-  description = "Cluster ID where the Karpenter controller is provisioned/managing"
+  description = "[Deprecated - use `karpenter_controller_cluster_name`] The name of the cluster where the Karpenter controller is provisioned/managing"
+  type        = string
+  default     = "*"
+}
+
+variable "karpenter_controller_cluster_name" {
+  description = "The name of the cluster where the Karpenter controller is provisioned/managing"
   type        = string
   default     = "*"
 }
@@ -234,6 +265,12 @@ variable "attach_load_balancer_controller_targetgroup_binding_only_policy" {
   description = "Determines whether to attach the Load Balancer Controller policy for the TargetGroupBinding only"
   type        = bool
   default     = false
+}
+
+variable "load_balancer_controller_targetgroup_arns" {
+  description = "List of Target groups ARNs using Load Balancer Controller"
+  type        = list(string)
+  default     = ["arn:aws:elasticloadbalancing:*:*:targetgroup/*/*"]
 }
 
 # AWS Appmesh Controller